initial refactor, with golang file structure
