+++ go.mod
@@ -1,8 +1,8 @@
-module github.com/digitallysavvy/agora-token-server
+module github.com/AgoraIO-Community/agora-token-server
 
 go 1.14
 
 require (
 	github.com/AgoraIO-Community/go-tokenbuilder v1.0.0
-	github.com/gin-gonic/gin v1.6.3
+	github.com/gin-gonic/gin v1.7.2
 )
