--- main.go
@@ -1,236 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"log"
-	"os"
-	"strconv"
-	"time"
-
-	"github.com/AgoraIO-Community/go-tokenbuilder/rtctokenbuilder"
-	"github.com/AgoraIO-Community/go-tokenbuilder/rtmtokenbuilder"
-	"github.com/gin-gonic/gin"
-)
-
-var appID string
-var appCertificate string
-
-func main() {
-
-	appIDEnv, appIDExists := os.LookupEnv("APP_ID")
-	appCertEnv, appCertExists := os.LookupEnv("APP_CERTIFICATE")
-
-	if !appIDExists || !appCertExists {
-		log.Fatal("FATAL ERROR: ENV not properly configured, check appID and appCertificate")
-	} else {
-		appID = appIDEnv
-		appCertificate = appCertEnv
-	}
-
-	api := gin.Default()
-
-	api.GET("/ping", func(c *gin.Context) {
-		c.JSON(200, gin.H{
-			"message": "pong",
-		})
-	})
-
-	api.Use(nocache())
-	api.GET("rtc/:channelName/:role/:tokentype/:uid/", getRtcToken)
-	api.GET("rtm/:uid/", getRtmToken)
-	api.GET("rte/:channelName/:role/:tokentype/:uid/", getBothTokens)
-	api.Run(":8080") // listen and serve on localhost:8080
-}
-
-func nocache() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		// set headers
-		c.Header("Cache-Control", "private, no-cache, no-store, must-revalidate")
-		c.Header("Expires", "-1")
-		c.Header("Pragma", "no-cache")
-		c.Header("Access-Control-Allow-Origin", "*")
-	}
-}
-
-func getRtcToken(c *gin.Context) {
-	log.Printf("rtc token\n")
-	// get param values
-	channelName, tokentype, uidStr, role, expireTimestamp, err := parseRtcParams(c)
-
-	if err != nil {
-		c.Error(err)
-		c.AbortWithStatusJSON(400, gin.H{
-			"message": "Error Generating RTC token: " + err.Error(),
-			"status":  400,
-		})
-		return
-	}
-
-	rtcToken, tokenErr := generateRtcToken(channelName, uidStr, tokentype, role, expireTimestamp)
-
-	if tokenErr != nil {
-		log.Println(tokenErr) // token failed to generate
-		c.Error(tokenErr)
-		errMsg := "Error Generating RTC token - " + tokenErr.Error()
-		c.AbortWithStatusJSON(400, gin.H{
-			"status": 400,
-			"error":  errMsg,
-		})
-	} else {
-		log.Println("RTC Token generated")
-		c.JSON(200, gin.H{
-			"rtcToken": rtcToken,
-		})
-	}
-}
-
-func getRtmToken(c *gin.Context) {
-	log.Printf("rtm token\n")
-	// get param values
-	uidStr, expireTimestamp, err := parseRtmParams(c)
-
-	if err != nil {
-		c.Error(err)
-		c.AbortWithStatusJSON(400, gin.H{
-			"message": "Error Generating RTC token: " + err.Error(),
-			"status":  400,
-		})
-		return
-	}
-
-	rtmToken, tokenErr := rtmtokenbuilder.BuildToken(appID, appCertificate, uidStr, rtmtokenbuilder.RoleRtmUser, expireTimestamp)
-
-	if tokenErr != nil {
-		log.Println(tokenErr) // token failed to generate
-		c.Error(tokenErr)
-		errMsg := "Error Generating RTM token: " + tokenErr.Error()
-		c.AbortWithStatusJSON(400, gin.H{
-			"error":  errMsg,
-			"status": 400,
-		})
-	} else {
-		log.Println("RTM Token generated")
-		c.JSON(200, gin.H{
-			"rtmToken": rtmToken,
-		})
-	}
-}
-
-func getBothTokens(c *gin.Context) {
-	log.Printf("dual token\n")
-	// get rtc param values
-	channelName, tokentype, uidStr, role, expireTimestamp, rtcParamErr := parseRtcParams(c)
-
-	if rtcParamErr != nil {
-		c.Error(rtcParamErr)
-		c.AbortWithStatusJSON(400, gin.H{
-			"message": "Error Generating RTC token: " + rtcParamErr.Error(),
-			"status":  400,
-		})
-		return
-	}
-	// generate the rtcToken
-	rtcToken, rtcTokenErr := generateRtcToken(channelName, uidStr, tokentype, role, expireTimestamp)
-	// generate rtmToken
-	rtmToken, rtmTokenErr := rtmtokenbuilder.BuildToken(appID, appCertificate, uidStr, rtmtokenbuilder.RoleRtmUser, expireTimestamp)
-
-	if rtcTokenErr != nil {
-		log.Println(rtcTokenErr) // token failed to generate
-		c.Error(rtcTokenErr)
-		errMsg := "Error Generating RTC token - " + rtcTokenErr.Error()
-		c.AbortWithStatusJSON(400, gin.H{
-			"status": 400,
-			"error":  errMsg,
-		})
-	} else if rtmTokenErr != nil {
-		log.Println(rtmTokenErr) // token failed to generate
-		c.Error(rtmTokenErr)
-		errMsg := "Error Generating RTC token - " + rtmTokenErr.Error()
-		c.AbortWithStatusJSON(400, gin.H{
-			"status": 400,
-			"error":  errMsg,
-		})
-	} else {
-		log.Println("RTC Token generated")
-		c.JSON(200, gin.H{
-			"rtcToken": rtcToken,
-			"rtmToken": rtmToken,
-		})
-	}
-
-}
-
-func parseRtcParams(c *gin.Context) (channelName, tokentype, uidStr string, role rtctokenbuilder.Role, expireTimestamp uint32, err error) {
-	// get param values
-	channelName = c.Param("channelName")
-	roleStr := c.Param("role")
-	tokentype = c.Param("tokentype")
-	uidStr = c.Param("uid")
-	expireTime := c.DefaultQuery("expiry", "3600")
-
-	if roleStr == "publisher" {
-		role = rtctokenbuilder.RolePublisher
-	} else {
-		role = rtctokenbuilder.RoleSubscriber
-	}
-
-	expireTime64, parseErr := strconv.ParseUint(expireTime, 10, 64)
-	if parseErr != nil {
-		// if string conversion fails return an error
-		err = fmt.Errorf("failed to parse expireTime: %s, causing error: %s", expireTime, parseErr)
-	}
-
-	// set timestamps
-	expireTimeInSeconds := uint32(expireTime64)
-	currentTimestamp := uint32(time.Now().UTC().Unix())
-	expireTimestamp = currentTimestamp + expireTimeInSeconds
-
-	return channelName, tokentype, uidStr, role, expireTimestamp, err
-}
-
-func parseRtmParams(c *gin.Context) (uidStr string, expireTimestamp uint32, err error) {
-	// get param values
-	uidStr = c.Param("uid")
-	expireTime := c.DefaultQuery("expiry", "3600")
-
-	expireTime64, parseErr := strconv.ParseUint(expireTime, 10, 64)
-	if parseErr != nil {
-		// if string conversion fails return an error
-		err = fmt.Errorf("failed to parse expireTime: %s, causing error: %s", expireTime, parseErr)
-	}
-
-	// set timestamps
-	expireTimeInSeconds := uint32(expireTime64)
-	currentTimestamp := uint32(time.Now().UTC().Unix())
-	expireTimestamp = currentTimestamp + expireTimeInSeconds
-
-	// check if string conversion fails
-	return uidStr, expireTimestamp, err
-}
-
-func generateRtcToken(channelName, uidStr, tokentype string, role rtctokenbuilder.Role, expireTimestamp uint32) (rtcToken string, err error) {
-
-	if tokentype == "userAccount" {
-		log.Printf("Building Token with userAccount: %s\n", uidStr)
-		rtcToken, err = rtctokenbuilder.BuildTokenWithUserAccount(appID, appCertificate, channelName, uidStr, role, expireTimestamp)
-		return rtcToken, err
-
-	} else if tokentype == "uid" {
-		uid64, parseErr := strconv.ParseUint(uidStr, 10, 64)
-		// check if conversion fails
-		if parseErr != nil {
-			err = fmt.Errorf("failed to parse uidStr: %s, to uint causing error: %s", uidStr, parseErr)
-			return "", err
-		}
-
-		uid := uint32(uid64) // convert uid from uint64 to uint 32
-		log.Printf("Building Token with uid: %d\n", uid)
-		rtcToken, err = rtctokenbuilder.BuildTokenWithUID(appID, appCertificate, channelName, uid, role, expireTimestamp)
-		return rtcToken, err
-
-	} else {
-		err = fmt.Errorf("failed to generate RTC token for Unknown Tokentype: %s", tokentype)
-		log.Println(err)
-		return "", err
-	}
-}
